+++ Frontend/src/app/(home)/folders/[folder]/page.tsx
@@ -1,8 +1,8 @@
-import FolderEmailList from '@modules/folders/folderEmailList';
+import FolderEmailListWrapper from "@features/folders/pages/email-folder-list";
 
 const page = async ({ params }: { params: Promise<{ folder: string }> }) => {
     const { folder } = await params;
-    return <FolderEmailList folderId={folder} />;
+    return <FolderEmailListWrapper folderId={folder} />;
 };
 
 export default page;