+++ Frontend/src/features/settings/api/settings.queries.ts
@@ -1,0 +1,12 @@
+import { QUERY_KEYS } from '@shared/api';
+import { APIResponse, ProfileSettingsDataObject } from '@shared/types';
+import { useQuery } from '@tanstack/react-query';
+import { getUserProfileSettings } from './settings.api';
+
+export const useGetProfileQuery = (enabled: boolean) => {
+    return useQuery<APIResponse<ProfileSettingsDataObject>>({
+        queryKey: [QUERY_KEYS.USER_PROFILE_SETTINGS],
+        queryFn: () => getUserProfileSettings(),
+        enabled,
+    });
+};