+++ Frontend/src/shared/components/inputs/TextInput.tsx
@@ -2,9 +2,9 @@
 
 import React from 'react';
 
+import { cn } from '@lib/utils';
 import { Input } from '@shared/ui/input';
 import { Label } from '@shared/ui/label';
-import { cn } from '../../../lib/utils';
 
 interface TextInputProps {
     label: string;